Modified vexload.cpp to fix problem where Mk6 labels had zeroes turned into Ohs.
There is a routine called fixOhs that takes the VSN and is supposed to
change any zeroes in the part of the label before the delimiting +, -
or % and turn them into capital Ohs.  The problem was that it didn't
have logic to recognize the percent sign as a delimiter and turned
zeroes in the numeric part of the VSN to Ohs as well.



git-svn-id: https://svn.atnf.csiro.au/difx@8217 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876