Update in difx2fits
  New option is added

--all-pcal-tones -A                  Extract all phase calibration tones

  This options overrides tone selection either specified in v2d file 
o set by default in embedded in the difx input files and extracts 
all the tines within recorded bandwidth. There are two side-effects.
If the phase calibration tone spacing is not commensurate to the IF width,
a situation may happen that the number of tones in different IFs may differ
by 1. When -A is applied the maximum number of tones is put in the 
FITS-IDI. Unused slots for phase calibration frequency, phase and
amplitude are padded with zeroes. Without -A option IFs with the same 
frequency range but with different phase calibration tones are considered
different. With -A option such IFs are considered identical.


git-svn-id: https://svn.atnf.csiro.au/difx@9728 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876