This version passes a regression with py2
There is something weird going on with the improperly named doSolve
parameter (CASA 4.7.2 thinks it's getting a boolean) and disabling
validation in the generatd polconvert_cli.py avoids the issue.

And it turns out Europeans with accents are screwed in py2 without
taking pains to specify encodings everywhere.  Removing accents for now.



git-svn-id: https://svn.atnf.csiro.au/difx@9771 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876